support more fields in circuits
Makes various structures generic over `PrimeField`. Adds a test vector
for parsing (though not evaluating) a circuit from longfellow that
implements a MAC.

This circuit's serialization is about 4.5 MB, so instead of storing it
inline in JSON, we now keep encoded circuits in binary files next to the
JSON test vector.